requirements.txt
langchain>=0.1.17
langchain-groq>=0.1.0
pandas>=1.5.0
openpyxl
python-dotenv
